resources/app_usage_dashboard/requirements.txt
janome==0.5.0
pandas==2.2.3
plotly==5.24.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-i18n==0.3.9
PyYAML==6.0.2
requests==2.32.2
scikit-learn==1.6.1
streamlit==1.44.1
tenacity==8.2.3
wordcloud==1.9.4 